RequestTelemetry: fill the ResponseCode on failure
Previously, the ResponseCode property of the Telemetry item was only filled when the telemetry was a success.

This brought a very subtle and hard to diagnose scenario where if you create a custom RequestTelemetry, and set it to failed, you we forced to also fill the ResponseCode to something not-empty otherwise your telemetry item would be silently dropped and would never show up in Application Insights.

This is a very easy mistake to make and the diagnostic of it was very tricky, pretty much forcing you to fall on the right stackoverflow answer to find out about it:
https://stackoverflow.com/questions/62500162/application-insights-requesttelemetry-not-showing-up-in-requests-after-exception

Examples of incorrect uses are also present in various places, including official Application Insights documentation:
https://docs.microsoft.com/en-us/azure/azure-monitor/app/custom-operations-tracking#http-request-in-owin-self-hosted-app
Notice the `.Success = false;` assignment with no ResponseCode being set, which causes the telemetry to be dropped.

I suggest that we always fill the property even for failures scenarios to ensure a correct default behavior for every users. I picked `500` as the value to match with the `200` for success which usually represent "Unknown error" on the HTTP side. I'm open to use anything else as long as it's not empty